refactor(系统设置): 服务集成字段替换联调
